% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_data.R
\docType{data}
\name{car_data}
\alias{car_data}
\title{Car data}
\format{
Car data
}
\source{
https://rpubs.com/harshaash/EFA
}
\usage{
car_data
}
\description{
This dataset contains 90 responses for 14 different variables that customers consider while purchasing car. The survey questions were framed using 5 point likert scale with 1 being very low and 5 being very high.
The data can be downloaded from the given link. The variables are the following:
1. Price
2. Safety
3. Exterior looks
4. Space and comfort
5. Technology
6. After sales service
7. Resale value
8. Fuel type
9. Fuel efficiency
10. Color
11. Maintenance
12. Test drive
13. Product reviews
14. Testimonials
}
\keyword{datasets}
