% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfnb.R
\name{dfnb}
\alias{dfnb}
\title{Discrete factor analysis for the negative binomial distribution}
\usage{
dfnb(y)
}
\arguments{
\item{y}{Data, an n by d numeric matrix}
}
\value{
A list with entries
\item{AIC}{AIC value for the optimal model}
\item{indexmat}{Factors and variables in each factor}
\item{estr0}{Estimated value of r for the negative binomial distributed factor(s)}
\item{estp0}{Estimated value of p for the negative binomial distributed factor(s)}
\item{estr}{Estimated value of r for the negative binomial distributed observations(s)}
\item{estp}{Estimated value of r for the negative binomial distributed observations(s)}
}
\description{
Discrete factor analysis for the negative binomial distribution
}
\examples{
dfnb(zinb100_Data[1:40,1:5])
}
