% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfp.R
\name{dfp}
\alias{dfp}
\title{Discrete factor analysis with the Poisson distribution}
\usage{
dfp(y)
}
\arguments{
\item{y}{Data, an n by d numeric matrix}
}
\value{
A list with entries
\item{AIC}{AIC value for the optimal model}
\item{indexmat}{Factors and variables in each factor}
\item{estlambda}{Estimated parameters for factors}
\item{estmu}{Estimated parameters for each variable within each factor}
}
\description{
Discrete factor analysis with the Poisson distribution
}
\examples{
dfp(car_data[,2:9])
}
