% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfzip.R
\name{dfzip}
\alias{dfzip}
\title{Discrete factor analysis with the zero inflated Poisson distribution}
\usage{
dfzip(y)
}
\arguments{
\item{y}{Data, an n by d numeric matrix}
}
\value{
A list with entries
\item{AIC}{AIC value for the optimal model}
\item{indexmat}{Factors and variables in each factor}
\item{estpilam}{Estimated zero-inflated parameters for  for each factor}
\item{estlam}{Estimated parameters for each factor}
\item{estpimu}{Estimated zero-inflated parameters for each variable within each factor}
\item{estmu}{Estimated parameters for each variable within each factor}
}
\description{
Discrete factor analysis with the zero inflated Poisson distribution
}
\examples{
dfzip(car_data[,1:3])
}
