% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_potato_data.R
\docType{data}
\name{original_potato_data}
\alias{original_potato_data}
\title{Original potato data}
\format{
Original potato data
}
\source{
https://cran.r-project.org/web/packages/agridat/index.html
}
\usage{
original_potato_data
}
\description{
This dataset contains 1152 observation with four variables that are either ordinal or counts. These variables are as follows:
x1: Energy factor in the ordinal scale of 1 to 2
x2: weight in the ordinal scale of 1 to 3
x3: damage category in the ordinal scale of 1 to 4
x4: the count of tubers in each combination of categories as an integer value.
}
\keyword{datasets}
