% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Survey data}
\format{
Survey data
}
\source{
http://openpsychometrics.org/_rawdata/
}
\usage{
survey_data
}
\description{
The questionnaire is five Likert scale data (strongly disagree (1) to strongly agree (5)).
For some negative questions, the scale was reversed. We choose 13 variables (sample size = 435).
Optimism Scale: opt1 to opt6.
Mastery Scale: mast1 to mast7 to identity the hidden factors.
}
\keyword{datasets}
