% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb100_Data.R
\docType{data}
\name{zinb100_Data}
\alias{zinb100_Data}
\title{zinb100 Simulated Data}
\format{
zinb100_Data
}
\usage{
zinb100_Data
}
\description{
100 Simulated samples from a  zero-inflated truncated (truncation at 6) negative binomial process with (1,3,4,2) factor structure.
}
\keyword{datasets}
